@@ -1,6 +1,6 @@
 {
 	"name": "skynovel",
-	"version": "3.15.4",
+	"version": "3.16.0",
 	"publisher": "famibee",
 	"displayName": "SKYNovel",
 	"description": "It supports development of SKYNovel for VSCode, VSCodeでSKYNovelの開発をサポートする拡張機能",
@@ -691,9 +691,9 @@
 		"@types/fs-extra": "^9.0.13",
 		"@types/glob": "^7.1.4",
 		"@types/mocha": "^9.0.0",
-		"@types/node": "^16.10.1",
+		"@types/node": "^16.10.2",
 		"@types/node-fetch": "^3.0.3",
-		"@types/power-assert": "^1.5.4",
+		"@types/power-assert": "^1.5.5",
 		"@types/uuid": "^8.3.1",
 		"@types/vscode": "^1.60.0",
 		"@typescript-eslint/eslint-plugin": "^4.32.0",
@@ -710,7 +710,7 @@
 		"typescript": "^4.4.3",
 		"vscode-debugadapter-testsupport": "^1.49.0",
 		"vscode-test": "^1.6.1",
-		"webpack": "^5.54.0",
+		"webpack": "^5.55.1",
 		"webpack-cli": "^4.8.0"
 	},
 	"optionalDependencies": {
