@@ -1,10 +1,35 @@
 // 別のタブにフォーカスして戻るたびに発生（丸ごと再生成？）するので、このような
 // リクエストして最新情報をホストにもらう、サーバー・クライアント方式になっている。
 // この事実を、createWebviewPanel()したホストは知りもしないので。
-// オブジェクト指向、メッセージドリブンと云えばそうなもかもしれんけど、さ。
 const vscode = ('acquireVsCodeApi' in window) ?acquireVsCodeApi() :null;
-window.addEventListener('message', e=> {
 //vscode.postMessage({cmd: 'info', text: 'setting.js'});	// デバッグ時はこう
+
+const hTemp = [{
+	'save_ns'		: 'hatsune',
+	'book.title'	: '初音館にて',
+	'book.creator'	: 'ふぁみべぇ',
+	'book.cre_url'	: 'https://twitter.com/famibee',
+	'book.publisher': '電子演劇部',
+	'book.pub_url'	: 'https://famibee.blog.fc2.com/',
+	'book.detail'	: '江戸川乱歩「孤島の鬼」二次創作ノベルゲームサンプルです。',
+}, {
+	'save_ns'		: 'uc',
+	'book.title'	: '桜の樹の下には',
+	'book.creator'	: 'ふぁみべぇ',
+	'book.cre_url'	: 'https://twitter.com/famibee',
+	'book.publisher': '電子演劇部',
+	'book.pub_url'	: 'https://famibee.blog.fc2.com/',
+	'book.detail'	: '梶井基次郎「桜の樹の下には」をノベルゲーム化したものです。',
+}];
+const chkVld = i=> {
+	if (! i.classList.contains('sn-vld')) return;
+
+	i.setCustomValidity('');
+	hTemp.forEach(v=> {if (v[i.id] === i.value) i.setCustomValidity('x');});
+	i.closest('form')?.checkValidity();
+};
+
+window.addEventListener('message', e=> {
 	if (! e.isTrusted) {
 		vscode.postMessage({cmd: 'warn', text: `(setting.js) isTrusted=false`});
 		return;
@@ -29,43 +54,10 @@ window.addEventListener('message', e=> {
 	elm.value = o.save_ns;
 	elm.focus();
 
-	const hTemp = [{
-		'save_ns'		: 'hatsune',
-		'book.title'	: '初音館にて',
-		'book.creator'	: 'ふぁみべぇ',
-		'book.cre_url'	: 'https://twitter.com/famibee',
-		'book.publisher': '電子演劇部',
-		'book.pub_url'	: 'https://famibee.blog.fc2.com/',
-		'book.detail'	: '江戸川乱歩「孤島の鬼」二次創作ノベルゲームサンプルです。',
-	}, {
-		'save_ns'		: 'uc',
-		'book.title'	: '桜の樹の下には',
-		'book.creator'	: 'ふぁみべぇ',
-		'book.cre_url'	: 'https://twitter.com/famibee',
-		'book.publisher': '電子演劇部',
-		'book.pub_url'	: 'https://famibee.blog.fc2.com/',
-		'book.detail'	: '梶井基次郎「桜の樹の下には」をノベルゲーム化したものです。',
-	}];
-	const chkEqTemp = i=> {
-		const cl = i.classList;
-		if (! cl.contains('sn-vld')) return;
-
-		const len = hTemp.length;
-		for (let j=0; j<len; ++j) {
-			if (hTemp[j][i.id] === i.value) {	// テンプレと同じ値は警告
-				cl.add('is-invalid');
-				cl.remove('is-valid');
-				break;
-			}
-
-			cl.add('is-valid');
-			cl.remove('is-invalid');
-		}
-	};
 	document.querySelectorAll('.sn-vld, .sn-gray').forEach(c=> {
-		chkEqTemp(c);
+		chkVld(c);
 		c.addEventListener('input', ()=> {
-			chkEqTemp(c);
+			chkVld(c);
 			vscode.postMessage({cmd: 'input', id: c.id, val: c.value});
 		}, {passive: true});
 	});
@@ -82,5 +74,18 @@ window.addEventListener('message', e=> {
 			});
 		}, {passive: true});
 	});
+
+	document.querySelectorAll('.form-range').forEach(c=> {
+		const rngV = c.closest('.range-wrap').querySelector('.range-badge');
+		const setValue = ()=> {
+			const	val = Number( (c.value - c.min) * 100 / (c.max - c.min) ),
+					pos = 10 - (val * 0.2);
+			rngV.innerHTML = `<span>${c.value}</span>`;
+			rngV.style.left = `calc(${val}% + (${pos}px))`;
+		};
+		setValue();
+		c.addEventListener('input', setValue, {passive: true});
+	});
+
 }, {passive: true});
 vscode?.postMessage({cmd: 'get'});
